add console.warn for user if jquery is not found on the page
This would have helped clear up some of my confusion earlier,
as I thought the jquery integration was required in the shim.

It also makes it possible to use the shim on pages without jquery
as it returns early if jquery is not found.  In our application
jquery is loaded on some pages, but not all.