Build process updates
Rollup is now only used for the browser bundle. Plain old `tsc` is used
to build our CommonJS and ES modules. TypeScript declarations are
generated per file automatically and included in the built module
directories. There will be more changes to come, but this provides a
good starting point to build from.