browser/src/routes: correctly index last element in responseType
In JavaScript we cannot reference array elements by providing a negative
index (unlike in Ruby or Python). The code incorrectly assumed that this works.
As result, the library could never send any response types to Airbrake.

The fix is simple: we reference the last element of the array the way JavaScript
wants.